Generator script now creates Section objects
The data is organised as follows. A Guide object holds the project
information, and contains a set of Section objects, each object holds
the information for a particular chapter/appendix/section of the guide.

Each Section will contain both the raw data, and the HTML output from
the parser.

TODO: Add function to number Section objects in order.