Added installation of dependencies
This is a very basic installation of dependencies using pip.
Dependencies are listed in dependencies.txt
The modules are not imported.

Pip's logging interferes with our own logging mechanisms and needs
to be shutdown before our logger begins. It also appears to show
our logging messages to standard output after being shutdown.