Text folder restructure
This commit includes a major restructure of the
text folder, which simplifies the creation
process by having the source text folder and
output folder to have same folder/file structure.

The files have also be renamed to have the
language abbreviation at the start of the file so
files can be easily organised in a file manager.