Move file loading to end of page (fixes #70)
Specific files that require loading at end of
page (due to size or jQuery) now load at the
end of the page. HTML for loading files are
appended to a list within the section object
and are passed through to Jinja2. Files that
require late loading in inpage interactives
are specificied by new tags.

To fix:
- Catch error if flags not found
- Ignore comments