Catch comments containing curly braces
This fixes #64 however it is not an ideal
solution. The regex is not very effective
and doesn't match gracefully but does the
job. The new regex will work with curly
braces within {comment}...{comment end}
blocks, but not within {comment} blocks.

It is probably worth consider removing
standalone {comment} blocks for
consistency in Markdown syntax and regex
matching.