Added custom colouring to panels (fixes #110)
I've added a very simple way to set colour
schemes of panels. This settings is stored
in the website.scss file as it alters the
page styling.

The CSS style is created from a SCSS mixin
(which is a small function) that takes
the panel type and colour.

Example:
@include panel-color("panel-curiosity", "green");

The colour should
be picked from the list available on
Materialize's website here:
http://materializecss.com/color.html#palette

Just the base colour is required (for example:
'red'), the mixin automatically picks shades
for the header, body, and arrow. A lighter shade
is used for the body to make it easier to read
longer sections of text.

I've picked colours for teacher note, curiosity,
and extra for expert panels to complement the
existing scheme.

Note: I've written the word 'colour' as 'color'
in the CSS files to remain consistent with
Materialize CSS.