Cleaned up the representing colours section
Most text was shortened, to reduce the overall wordiness, and some
concepts were moved to curiosity boxes as they were not essential
for the overall text.

The hexadecimal colour codes stuff was moved to a curiosity box, because
it is not actually central to representing images using bits, and is instead
just an additional extra that students might be able to relate to. Putting it
in the main text was putting too much weight on its relevance.