Show binary representation rather than decimal (fixes #285)
The interactive now shows the colour values in
binary and allows the user to click a bit to
change a colour value.

The JS code was programmed for proof of concept,
which can be rolled out to production if needed.
However the JS code should be reprogrammed at
a later date.