Removed incorrect function call (noted on #333)
This function is used to remove an object from a scene and therefore
clears it from memory.
I have deleted this function because it makes the cube disappear too
suddenly, and memory leaks should not be an issue since the interactive
is likely to be used for a short time only.