Fix detection of line number for missing type argument for panel
When a panel does not have a type provided
an error is logged, with the matched markdown
to find the original line number. The text
provided was the text for the whole panel,
rather than just the first line, so no line
every matched the text.

This has been fixed by only passing through
the first line of the matched markdown, and
fixes #377.