WIP fixing bug in translate and scale (for matrices).
Ok, turns out the bug is so much worse than I thought. The two matrix
operations are getting the way of each other, i.e. if you scale the
arrow then translate it, it changes it back to it's original shape.

Fixed the translate function however to make sure the translate matrix
changed, rather than translating it repeatedly with the same x,y values.

 Please enter the commit message for your changes. Lines starting