Add prototype of NCEA encoding selector interactive
The interactive allows students to select which
guides they are following for Level 2 NCEA encoding
and which grade it can provide them.

The basic HTML and CSS is here but the structure of
the table needs to be changed to be ordered with
topics across the top, rather than achievement level.

JS logic is currently missing too, and this will be
added next once structure is fixed.