Add dynamic creation and validation of ncea-encoding-selector items
The structure and max values are stored
within a dictionary structure, which is
used to create the initial grid.

While the JS requires looping through the
values for counting, and then again to
alter the grid (for disabling or re-enabling
items, or adding or removing validation
colours), it does work with different numbers
of items, grades, etc. This may allow this
interactive to be used for other NCEA course
selections.

Feedback messages are next to be added under
the grid.