Handle chapters in database properly. See desc
Why does this work? Previously, the slug was used as the key when
updating the sections, however if you added a new section in the middle
then you would end up with multiple entries with the same `number` as
the previous one wasn't deleted. This caused a key constraint
violation error. It makes more sense for the `number` to be the key
here, and to update the slug associated with that `number`. If a slug
is used multiple times or if a `number` is repeated, this will be a
domain constraint violation, an invalid sequence so we don't need to
worry about those cases here.